Extract mode module into separate file
